Move custom header components to components folder, and create exporting files
