+++ react/components/Layout.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Layout = ({ children }) => (
+  <React.Fragment>
+    {children}
+  </React.Fragment>
+)
+
+export default Layout
