+++ react/components/Spacer.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+const Spacer = () => <div className="flex flex-grow-1" />
+
+export default Spacer
